requirements.txt
langchain-google-genai
python-dotenv
langchain
langchain-core
streamlit